+++ EDDI/ChangeLog.md
@@ -2,6 +2,14 @@
 
 Full details of the variables available for each noted event, and VoiceAttack integrations, are available in the individual [event pages](https://github.com/EDCD/EDDI/wiki/Events).
 
+### Development
+  
+  * Speech responder
+    * Added new `Next jump` event
+    * Added new top level object `nextsystem` - like `lastsystem` but for the next system you are visiting. e.g. "Our next waypoint is \{nextsystem.name\}".
+  * VoiceAttack  
+    * Added new `Next system` variables
+    
 ### 3.3.4-rc3
   * Frontier API
     * Fixed missing client ID in 3.3.4-rc2.